chore: move all the end to end tests to one folder
